refactor playlist test helpers, parametrize monkeypatch
Because:
- Reduce code duplication in playlist test setup
- Enable easier monkeypatching and content generation

Moves WriteSpy and playlist content generation to helpers.py, updates all playlist_factory usages to inject monkeypatch, and centralizes content logic for maintainability.
